pyproject.toml
[tool.poetry]
name = "python-clean-arch"
version = "0.1.0"
description = ""
authors = ["Robin Shi <robinshi@outlook.com>"]
readme = "README.md"
packages = [{ include = "python_clean_arch" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.19"
pydantic = "^2.0.3"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.2"
uvicorn = { extras = ["standard"], version = "^0.23.1" }
dependency-injector = "^4.41.0"
strawberry-graphql = { extras = ["debug-server"], version = "^0.195.3" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
